+++ raiden-ts/src/raiden.ts
@@ -71,6 +71,7 @@ import { RaidenError, ErrorCodes } from './utils/error';
 export class Raiden {
   private readonly store: Store<RaidenState, RaidenAction>;
   private readonly deps: RaidenEpicDeps;
+  public config: RaidenConfig;
 
   /**
    * action$ exposes the internal events pipeline. It's intended for debugging, and its interface
@@ -232,6 +233,9 @@ export class Raiden {
       },
     });
 
+    this.config = { ...defaultConfig, ...state.config };
+    this.deps.config$.subscribe((config) => (this.config = config));
+
     this.deps.config$
       .pipe(pluckDistinct('logger'))
       .subscribe((logger) => this.log.setLevel(logger || 'silent', false));
@@ -258,6 +262,7 @@ export class Raiden {
    * An async factory is needed so we can do the needed async requests to construct the required
    * parameters ahead of construction time, and avoid partial initialization then
    *
+   * @param this - Raiden class or subclass
    * @param connection - A URL or provider to connect to, one of:
    *     <ul>
    *       <li>JsonRpcProvider instance,</li>
@@ -279,8 +284,9 @@ export class Raiden {
    * @param config - Raiden configuration
    * @param subkey - Whether to use a derived subkey or not
    * @returns Promise to Raiden SDK client instance
-   **/
-  public static async create(
+   */
+  public static async create<R extends typeof Raiden>(
+    this: R,
     connection: JsonRpcProvider | AsyncSendable | string,
     account: Signer | string | number,
     storageOrState?:
@@ -291,7 +297,7 @@ export class Raiden {
     contracts?: ContractsInfo,
     config?: PartialRaidenConfig,
     subkey?: true,
-  ): Promise<Raiden> {
+  ): Promise<InstanceType<R>> {
     let provider: JsonRpcProvider;
     if (typeof connection === 'string') {
       provider = new JsonRpcProvider(connection);
@@ -332,7 +338,15 @@ export class Raiden {
       `Mismatch between network or registry address and loaded state`,
     );
 
-    const raiden = new Raiden(provider, network, signer, contracts, state, defaultConfig, main);
+    const raiden = new this(
+      provider,
+      network,
+      signer,
+      contracts,
+      state,
+      defaultConfig,
+      main,
+    ) as InstanceType<R>;
     if (onState) raiden.state$.subscribe(onState, onStateComplete, onStateComplete);
     return raiden;
   }
@@ -416,17 +430,6 @@ export class Raiden {
    */
   public async getBlockNumber(): Promise<number> {
     return this.deps.provider.blockNumber || (await this.deps.provider.getBlockNumber());
-  }
-
-  /**
-   * Getter for current Raiden Config
-   *
-   * @returns Current Raiden config
-   */
-  public get config(): RaidenConfig {
-    let config!: RaidenConfig;
-    this.deps.config$.pipe(first()).subscribe((c) => (config = c));
-    return config;
   }
 
   /**